+++ internal/updater/launch_linux_helper.go
@@ -1,0 +1,7 @@
+//go:build linux
+
+package updater
+
+func HandleUpdateHelper([]string) (bool, error) {
+	return false, nil
+}