+++ internal/updater/launch_other.go
@@ -7,6 +7,10 @@ import (
 	"runtime"
 )
 
-func LaunchInstaller(string) error {
+func LaunchInstaller(string, string) error {
 	return fmt.Errorf("automatic updates do not support %s", runtime.GOOS)
+}
+
+func HandleUpdateHelper([]string) (bool, error) {
+	return false, nil
 }